Move service hooks file to a helper folder
